code crafters/requirements.txt
streamlit
mysql-connector-python
werkzeug
pandas
plotly
python-dotenv

